+++ src/main/java/com/ideality/coreflow/project/query/service/impl/DeptQueryServiceImpl.java
@@ -1,8 +1,8 @@
-package com.ideality.coreflow.holiday.query.service.impl;
+package com.ideality.coreflow.project.query.service.impl;
 
 import com.ideality.coreflow.common.exception.BaseException;
-import com.ideality.coreflow.holiday.query.mapper.DeptMapper;
-import com.ideality.coreflow.holiday.query.service.DeptQueryService;
+import com.ideality.coreflow.project.query.mapper.DeptMapper;
+import com.ideality.coreflow.project.query.service.DeptQueryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
